Move selectedTransactions to Search Context
