+++ src/pages/Search/SearchHoldReasonPage.tsx
@@ -25,11 +25,13 @@ type SearchHoldReasonPageProps = {
 function SearchHoldReasonPage({route}: SearchHoldReasonPageProps) {
     const {translate} = useLocalize();
 
-    const {currentSearchHash, selectedTransactionIDs} = useSearchContext();
+    const {currentSearchHash, selectedTransactions, clearSelectedTransactions} = useSearchContext();
     const {backTo = ''} = route.params ?? {};
 
+    const selectedTransactionIDs = Object.keys(selectedTransactions);
     const onSubmit = (values: FormOnyxValues<typeof ONYXKEYS.FORMS.MONEY_REQUEST_HOLD_FORM>) => {
         SearchActions.holdMoneyRequestOnSearch(currentSearchHash, selectedTransactionIDs, values.comment);
+        clearSelectedTransactions();
         Navigation.goBack();
     };
